feat: (#740) 활성화 정보 포함 알림버튼을 활성화정보 포함 버튼 컨테이너로 수정
