+++ frontend/src/components/common/ActiveContainer/style.ts
