--- frontend/src/components/common/AlarmIconButton/index.tsx
@@ -1,17 +1,0 @@
-import IconButton from '../IconButton';
-
-import * as S from './style';
-
-interface AlarmIconButtonProps {
-  isActive: boolean;
-  handleAlarmOpenClick: () => void;
-}
-
-export default function AlarmIconButton({ isActive, handleAlarmOpenClick }: AlarmIconButtonProps) {
-  return (
-    <S.Container>
-      <S.Active $isActive={isActive} />
-      <IconButton category="alarm" onClick={handleAlarmOpenClick} />
-    </S.Container>
-  );
-}
